+++ lib/l10n/app_localizations_en.dart
@@ -31,7 +31,7 @@ class AppLocalizationsEn extends AppLocalizations {
 
   @override
   String get about_description =>
-      'I first wrote and released TravelRates in 2019 in Bali whilst being sick with a cold when backpacking south east Asia.\n\nIt was built with backpackers in mind, where you have to compare against multiple currencies and speed, ease and offline use are more important than being exactly right.\n\nLet me know if you enjoy the app, every review motivates me to keep it running (and free)!';
+      'I first wrote and released TravelRates in 2019 in Bali whilst being sick with a cold when backpacking south east Asia.\n\nIt was built with backpackers in mind, where you have to compare against multiple currencies and speed, ease and offline use are more important than being exactly right.\n\nThe the rates are updated from online source whenever you are online, and the online rates are updated once a day.\n\nLet me know if you enjoy the app, every review motivates me to keep it running (and free)!';
 
   @override
   String get about_addReview => 'Add review';
@@ -77,4 +77,22 @@ class AppLocalizationsEn extends AppLocalizations {
   @override
   String get addCurrency_digitalCurrency =>
       'Is a digital currency used worldwide';
+
+  @override
+  String get time_justNow => 'just now';
+
+  @override
+  String time_minutesAgo(int minutes) {
+    return '${minutes}m ago';
+  }
+
+  @override
+  String time_hoursAgo(int hours) {
+    return '${hours}h ago';
+  }
+
+  @override
+  String time_daysAgo(int days) {
+    return '${days}d ago';
+  }
 }